requirements.txt
streamlit
nltk
Pillow
tiktoken